contract: Change ContractBaseData#taxRate from double to BigDecimal
A disadvatange of using the primitive double was that is wasn't possible
to encode a missing value, causing the form to always show 0.0, even if
the value was not set. I decided to use BigDecimal over Double because
the ContractEntity also uses a BigDecimal, which makes mapping easier.
For the divisions the scale has been specified to prevent values from
being truncated when rounding. I chose 4 as the scale because that's
what we use in the database.

+++ budgeteer-web-interface/src/main/java/org/wickedsource/budgeteer/service/contract/ContractBaseData.java
@@ -10,6 +10,7 @@ import org.wickedsource.budgeteer.service.invoice.InvoiceBaseData;
 import org.wickedsource.budgeteer.web.components.fileUpload.FileUploadModel;
 
 import java.io.Serializable;
+import java.math.BigDecimal;
 import java.util.Date;
 import java.util.LinkedList;
 import java.util.List;
@@ -24,7 +25,7 @@ public class ContractBaseData implements Serializable{
     private Money budget;
     private Money budgetSpent;
     private Money budgetLeft;
-    private double taxRate; 
+    private BigDecimal taxRate;
     private String internalNumber;
     private Date startDate;
     private ContractEntity.ContractType type;
